changed webinstaller logic

+++ webinstaller.sh
@@ -29,22 +29,39 @@ BASHRC="${HOME}/.bashrc"
 SHELLFILES="${HOME}/.shellfiles"
 SCRIPTS="${HOME}/System32"
 
-################################# LOGIC ########################################
+################################# FUNCTIONS ######################################
 
-installScripts() {
+confirmation() {
+	local confirm
+	local question="${1}"
+
+	printf "%b" "${question}? (y/N) " && read -r confirm
+	printf '%b\n' ""
+
+	[[ "${confirm}" =~ ^[yY][eE]?[sS]?$ ]] && return 0
+
+	return 1
+}
+
+install_scripts() {
+	confirmation "Install scripts" || return 0
+
 	printf "%b\n" "Installing scripts..."
 	[[ -L "${SYM_SCRIPTS}" ]] || ln -sf "${SYM_SCRIPTS}" "${SCRIPTS}"
 }
 
-installShell() {
+install_shellfiles() {
+	confirmation "Install shell configs" || return 0
+
 	printf "%b\n" "Installing shell configs..."
 	[[ -L "${SYM_ZSHRC}" ]] || ln -sf "${SYM_ZSHRC}" "${ZSHRC}"
 	[[ -L "${SYM_BASHRC}" ]] || ln -sf "${SYM_BASHRC}" "${BASHRC}"
 	[[ -L "${SYM_SHELLFILES}" ]] || ln -sf "${SYM_SHELLFILES}" "${SHELLFILES}"
 }
 
-installConfigs() {
-	# Configs
+install_configs() {
+	confirmation "Install miscellaneous configs" || return 0
+
 	printf "%b\n" "Installing miscellaneous configs..."
 
 	for folder in "${DOTFILES}"/.config/*; do
@@ -57,35 +74,20 @@ installConfigs() {
 	done
 }
 
+################################# LOGIC ########################################
+
 # Check if Git is installed.
 [[ ! -x "$(command -v git)" ]] &&
 	printf '%b\n' "Git is not installed." && exit 1
 
-# Confirm whether to install dotfiles.
-printf '%b' "Are you sure you want to install this script?
-This script will delete certain files, including your current bashrc.\n
-Confirm? (y/N) " && read -r confirmInstall && printf "%b\n" ""
-
-[[ ! "${confirmInstall}" =~ ^[yY][eE]?[sS]?$ ]] &&
-	printf "%b\n" "Cancelling." && exit 1
+confirm "Are you sure you want to install this" || exit 1
 
 # Download dotfiles
 printf "%b\n" "Downloading dotfiles..."
 git pull "${DOTFILES_SITE}" "${DOTFILES_DOWN_DIR}"
 
-# Scripts
-printf "%b" "Install scripts? (y/N) " && read -r confirmScripts
-printf '%b\n' ""
-[[ "${confirmScripts}" =~ ^[yY][eE]?[sS]?$ ]] && installScripts
-
-# Shell configs
-printf "%b" "Install shell configs? (y/N) " && read -r confirmShell
-printf '%b\n' ""
-[[ "${confirmShell}" =~ ^[yY][eE]?[sS]?$ ]] && installShell
-
-# Misc configs
-printf "%b" "Install miscellaneous configs? (y/N) " && read -r confirmConfs
-printf '%b\n' ""
-[[ "${confirmConfs}" =~ ^[yY][eE]?[sS]?$ ]] && installConfigs
+install_scripts
+install_shellfiles
+install_configs
 
 printf "%b\n" "Done."
